// ionicqr-master/src/app/components/interfaces/enums/returnFormat.enum.ts
export enum ReturnFormat{
    yyyyMMdd_HHmmss = "yyyyMMdd_HHmmss",
    yyyyddMM_HHmmss = "yyyyddMM_HHmmss",
    MMddYYYY_HHmmss = "MMddYYYY_HHmmss",
    ddMMyyyy_HHmmss = "ddMMyyyy_HHmmss",
    MMdd_HHmmssm = "MMdd_HHmmssm",
    ddMMyyyy = "ddMMyyyy",
    HHmmss = "HHmmss",
    MMddyyyy = "MMddyyyy",
    yyyyMMdd = "yyyyMMdd",
    MMyyyy = "MMyyyy"
}